Upgrade core.async to be Clojure 1.9 compatible
Versions of core.async older than 0.3.442 are broken with Clojure 1.9-alpha15 and newer. Morse itself seems to work fine, this is just to make the dependencies work out of the box.

Meanwhile, if somebody is getting a spec exception about `:refer-clojure` when loading Morse, try adding `[org.clojure/core.async "0.3.465"]` as a dependency to your project.